Fix: Se mejoró la lógica de iniciar y finalizar viaje

+++ src/main/java/com/example/unirideapi/dto/request/RutaRequestDTO.java
@@ -3,6 +3,7 @@ package com.example.unirideapi.dto.request;
 import com.example.unirideapi.model.enums.EstadoRuta;
 import lombok.Builder;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.time.LocalTime;
 
@@ -12,7 +13,7 @@ public record RutaRequestDTO(
         String destino,
         LocalDate fechaSalida,
         LocalTime horaSalida,
-        Float tarifa,
+        BigDecimal tarifa,
         Integer asientosDisponibles,
         EstadoRuta estadoRuta,
         Integer conductorId
